Better representation of constraints in the builder IR
