+++ testdata/jennies/builders/constraints/schema.cue
@@ -1,4 +1,4 @@
-package sandbox
+package constraints
 
 import "strings"
